src/monitor/requirements.txt
paho-mqtt==1.6.1
influxdb-client>=1.38.0